Handle explicit import naming in mockery generator
This required two separate fixes:

The first of these was to lookup package aliases when making a call to
renderType so that references actually go through the type alias. The
second was to lookup the actual file within a package that contains the
interface definition so package imports are properly added in
GeneratePrologue.

Fixes: #47