Aggresively refactor import generation
The Generator now eagerly goes through all types used in methods of the
interface to be mocked and performs a lookup of to ensure that all
package import names are unique. This fixes support for imports that are
needed for methods coming from a nested interface whose names conflict
with imports coming from the main interface or other nested interfaces.

Fixes: #94